% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQcChecks.R
\name{runQCChecks}
\alias{runQCChecks}
\title{Runs quality checks on a dataset}
\usage{
runQCChecks(filename, datasetverID, recordType)
}
\arguments{
\item{filename}{the filename of the file to check}

\item{datasetverID}{the dataset ID}

\item{recordType}{string name of the record type}
}
\value{
The list of Datasets that can be screened in DATSU with the IDs
}
\description{
Check a data set against the ICES DATSU data submission utility,
the user must supply a dataset version and record type.
}
\examples{
\donttest{
filename <- system.file("test_files/vms_test.csv", package = "icesDatsuQC")
runQCChecks(filename, 145, "VE")
}
}
